+++ packages/atomic/src/components/commerce/index.ts
@@ -1,2 +1,2 @@
-export {AtomicCommerceLoadMoreProducts} from './atomic-commerce-load-more-products/atomic-commerce-load-more-products';
-export {AtomicComponentError} from '../common/atomic-component-error/atomic-component-error';
+// Auto-generated file
+export {AtomicCommerceLoadMoreProducts} from './atomic-commerce-load-more-products/atomic-commerce-load-more-products.js';