+++ packages/atomic/src/components/commerce/lazy-index.ts
@@ -1,10 +1,9 @@
+// Auto-generated file
 export default {
-  'atomic-commerce-load-more-products': () =>
-    import(
+  'atomic-commerce-load-more-products': async () =>
+    await import(
       './atomic-commerce-load-more-products/atomic-commerce-load-more-products.js'
     ),
-  'atomic-component-error': () =>
-    import('../common/atomic-component-error/atomic-component-error'),
 } as Record<string, () => Promise<unknown>>;
 
 export type * from './index.js';