+++ packages/atomic/src/components/common/interface/interface-common.tsx
@@ -3,7 +3,6 @@ import {LogLevel} from '@coveo/headless';
 import {i18n, TFunction} from 'i18next';
 import Backend from 'i18next-http-backend';
 import {html} from 'lit';
-import {loadFocusVisiblePolyfill} from '../../../global/focus-visible.js';
 import {loadDayjsLocale} from '../../../utils/dayjs-locales.js';
 import {AnyBindings, AnyEngineType} from './bindings.js';
 import {i18nBackendOptions, i18nTranslationNamespace} from './i18n.js';
@@ -42,7 +41,6 @@ export class CommonAtomicInterfaceHelper<Engine extends AnyEngineType> {
     globalVariableName: string
   ) {
     setCoveoGlobal(globalVariableName);
-    loadFocusVisiblePolyfill();
 
     if ('connectedCallback' in atomicInterface && 'render' in atomicInterface) {
       const {