+++ packages/atomic/src/components/insight/atomic-insight-numeric-facet/atomic-insight-numeric-facet.tsx
@@ -27,7 +27,7 @@ import {
 } from '../../../utils/initialization-utils';
 import {MapProp} from '../../../utils/props-utils';
 import {FocusTargetController} from '../../../utils/stencil-accessibility-utils';
-import {randomID} from '../../../utils/utils';
+import {randomID} from '../../../utils/stencil-utils';
 import {parseDependsOn} from '../../common/facets/depends-on';
 import {shouldDisplayInputForFacetRange} from '../../common/facets/facet-common';
 import {FacetInfo} from '../../common/facets/facet-common-store';