+++ packages/atomic/src/components/search/result-lists/atomic-result-children/atomic-result-children.tsx
@@ -18,7 +18,7 @@ import {
   InitializableComponent,
   InitializeBindings,
 } from '../../../../utils/initialization-utils';
-import {elementHasAncestorTag} from '../../../../utils/utils';
+import {elementHasAncestorTag} from '../../../../utils/stencil-utils';
 import {extractUnfoldedItem} from '../../../common/interface/item';
 import {
   ChildTemplatesContext,