+++ packages/atomic/src/components/search/smart-snippets/atomic-smart-snippet/atomic-smart-snippet.tsx
@@ -14,8 +14,8 @@ import {
   BindStateToController,
 } from '../../../../utils/initialization-utils';
 import {ArrayProp} from '../../../../utils/props-utils';
+import {randomID} from '../../../../utils/stencil-utils';
 import {shouldDisplayOnCurrentTab} from '../../../../utils/tab-utils';
-import {randomID} from '../../../../utils/utils';
 import {createAppLoadedListener} from '../../../common/interface/store';
 import {getAttributesFromLinkSlot} from '../../../common/item-link/attributes-slot';
 import {SmartSnippetCommon} from '../../../common/smart-snippets/atomic-smart-snippet/smart-snippet-common';