+++ packages/atomic/src/utils/initialization-utils.tsx
@@ -17,13 +17,7 @@ import {
   initializableElements,
   initializeEventName,
 } from './initialization-lit-stencil-common-utils';
-import {closest} from './utils';
-
-declare global {
-  interface Window {
-    applyFocusVisiblePolyfill?: (shadowRoot: ShadowRoot) => void;
-  }
-}
+import {closest} from './stencil-utils';
 
 export type InitializeEvent = CustomEvent<InitializeEventHandler>;
 
@@ -31,6 +25,7 @@ export type InitializeEvent = CustomEvent<InitializeEventHandler>;
  * Retrieves `Bindings` or `CommerceBindings` on a configured parent interface.
  * @param event - The element on which to dispatch the event, which must be the child of a configured Atomic container element.
  * @returns A promise that resolves upon initialization of the parent container element, and rejects otherwise.
+ * @deprecated should only be used for Stencil components. For Lit components, use `initializeBindings` from @/src/decorators/initialize-bindings.
  */
 export function initializeBindings<
   SpecificBindings extends AnyBindings = Bindings,
@@ -78,27 +73,6 @@ export interface InitializableComponent<
    */
   initialize?: () => void;
   error: Error;
-}
-
-/**
- * Makes Shadow Dom elements compatible with the focus-visible polyfill https://github.com/WICG/focus-visible
- * Necessary for Safari under version 15.4.
- */
-export function applyFocusVisiblePolyfill(element: HTMLElement) {
-  if (!element.shadowRoot) {
-    return;
-  }
-
-  if (window.applyFocusVisiblePolyfill) {
-    window.applyFocusVisiblePolyfill(element.shadowRoot);
-    return;
-  }
-
-  window.addEventListener(
-    'focus-visible-polyfill-ready',
-    () => window.applyFocusVisiblePolyfill?.(element.shadowRoot!),
-    {once: true}
-  );
 }
 
 type InitializeBindingsProps = {
@@ -228,7 +202,6 @@ export function InitializeBindings<SpecificBindings extends AnyBindings>({
       componentDidRender && componentDidRender.call(this);
       if (element.getAttribute(loadedAttribute) === 'false') {
         element.setAttribute(loadedAttribute, 'true');
-        applyFocusVisiblePolyfill(getElement(this));
         componentDidLoad && componentDidLoad.call(this);
       }
     };