+++ packages/atomic/src/utils/resource-url.ts
@@ -1,6 +1,16 @@
-/// <reference types="../types/custom-env" />
-// TODO: get rid of the mock once no longer running stencil tests
+const cdnURLs = [
+  'https://staticdev.cloud.coveo.com',
+  'https://staticstg.cloud.coveo.com',
+  'https://static.cloud.coveo.com',
+];
 
-// import.meta is an ESM only feature and cannot be used in CJS.
-// To avoid errors such as "SyntaxError: Cannot use 'import.meta' outside a module error" when running stencil tests, we need to mock this function in the tests.
-export const getResourceUrl = () => import.meta.env?.RESOURCE_URL;
+function isCoveoCDN() {
+  return cdnURLs.includes(new URL(import.meta.url).origin);
+}
+
+function getCoveoCdnResourceUrl() {
+  return import.meta.resolve('../../');
+}
+
+export const getResourceUrl = () =>
+  isCoveoCDN() ? getCoveoCdnResourceUrl() : undefined;