+++ packages/samples/headless-ssr-commerce/components/standalone-search-box.tsx
@@ -26,6 +26,9 @@ export default function StandaloneSearchBox() {
       const url = `${state.redirectTo}?q=${encodeURIComponent(state.value)}`;
       router.push(url, {scroll: false});
       methods?.afterRedirection();
+    } else if (state.redirectTo !== '') {
+      // This handles query pipeline redirect triggers.
+      window.location.replace(state.redirectTo);
     }
   }, [state.redirectTo, state.value, router, methods]);